fixed issue with moving of ExternalLinks component
