+++ .vscode/settings.json
@@ -1,10 +1,21 @@
 {
   "cSpell.words": [
     "alecdowning",
+    "Amazondynamodb",
+    "Amazonwebservices",
     "Appstore",
+    "Awslambda",
+    "Chakraui",
     "embla",
+    "Googleadmob",
+    "Handlebarsdotjs",
     "headshot",
     "jost",
-    "Partykit"
+    "Nextdotjs",
+    "Nodedotjs",
+    "Openapiinitiative",
+    "Partykit",
+    "Radixui",
+    "Reactrouter"
   ]
 }
