+++ src/components/ExternalLinks/index.ts
@@ -1,0 +1,1 @@
+export { ExternalLinks } from "./ExternalLinks";
