+++ src/components/index.ts
@@ -5,5 +5,6 @@ export { Header } from "./Header";
 export { ProjectsSection, projects, FullProject } from "./Projects";
 export { Button } from "./ui/button";
 export { StackDisplay } from "./StackDisplay";
+export { ExternalLinks } from "./ExternalLinks";
 
 export * from "./ui";
