+++ src/pages/Home/sections/Hero.tsx
@@ -1,35 +1,5 @@
-import { FaGithubSquare, FaLinkedin, FaEnvelope } from "react-icons/fa";
-import { StackDisplay } from "./StackDisplay";
+import { StackDisplay, ExternalLinks } from "@/components";
 
-export const ExternalLinks = () => (
-  <div className="w-full md:w-auto flex md:flex-col items-end justify-around md:justify-end gap-4">
-    <a
-      href="https://www.linkedin.com/in/alec-downing"
-      target="_blank"
-      className=" text-accent"
-      rel="noreferrer"
-    >
-      <FaLinkedin className="w-12 h-12 mt-[3px] text-brand-400/90 hover:scale-105 active:text-brand-400 transition-all" />
-    </a>
-
-    <a href="https://www.github.com/a-down" target="_blank" rel="noreferrer">
-      <FaGithubSquare className="w-12 h-12 mt-[3px]  text-brand-400/90 hover:scale-105 active:text-brand-400 transition-all" />
-    </a>
-
-    <a
-      href="mailto:alec@alecdowning.com"
-      className="flex items-center gap-4 text-brand-400/90 hover:scale-105 active:text-brand-400 transition-all"
-    >
-      {/* <p className="hidden lg:inline">alec@alecdowning.com</p> */}
-      <FaEnvelope className="w-11 h-12 mr-[3px]" />
-    </a>
-  </div>
-);
-
-/**
- *
- * @returns {Element}
- */
 export function Hero() {
   const heroTechStack = [
     "TypeScript",
